web/requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
braintree==4.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.5
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.7
djangorestframework==3.11.0
future==0.18.2
h11==0.9.0
html5lib==1.0.1
httptools==0.1.1
idna==2.9
kombu==4.6.10
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
Pyphen==0.9.5
pytz==2020.1
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
websockets==8.1
